backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
openai==1.51.0
anthropic==0.34.2
python-dotenv==1.0.1
sse-starlette==2.1.3
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
